Fix verification in HTTP library.
Clients get a list of public keys to verify the signed token, the client
library will try all the keys one by one until it find a match. All
these keys used to have the same size, 1024 bit.

Recently we've made all the newly generated key with 2048 bit, so it's
possible that a client get a list of two keys with the first key 1024
bit RSA and the second key 2048 bit RSA.

If the blob was signed by the 1024 bit key, and the client tried the
2048 bit key first, an exception will be throw, and vice versa.

The correct behavior is to ignore this exception and continue to try the
next key.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=95790730