Fixed bug where Special Values of Float and Double were not parsed from String and threw an exception.
The bug consists in the fact that NaN / Infinity / -Infinity are
interpreted as a String (which it is). I added a check that if the
expected Class is Float / Double (or float / double) and getText()
returns NaN / Infinity / -Infinity (case-insensitive), the Parses omits
checking that there is no JsonString Annotation.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=101010886