Preserve lowLevelHttpRequest, so that it's available to tests.
When lowLevelHttpRequest is not set via MockHttpTransport.Builder, buildRequest() creates a new instance of MockLowLevelHttpTransport. Due to what seems to be a bug, the newly created MLLHT wasn't saved to MHT instance. That means, getLowLevelHttpRequest() called on such instance returns null, making it impossible to inspect actual request being sent. For example, this bug made it impossible to verify headers of outbound HTTP requests.