base64: strip carriage returns and newlines
The guava [1] doc mentions that carriage returns and newlines are not
permitted, and recommend using .withSeparator. We do not know whether
the input string will contain a carriage return, or a newline, or both,
so instead we'll just strip them each individually.

This fixes an internal google bug that was introduced in as part of
the move from Apache Commons Codec to Guava in
ef9f9142be5dacdd19a3a16f37000eca478d2d51.

1: https://google.github.io/guava/releases/19.0/api/docs/com/google/common/io/BaseEncoding.html#withSeparator(java.lang.String,%20int)