Only encode the streaming content once.
In order to calculate the contentLength, we must write the encode the
data first. The encoded data is written to a buffer using a
ByteArrayOutputStream implementation and we use that to figure out how
many bytes of data is to be sent.

Previously, this data was thrown away and the content was re-encoded
when it was actually time to send the data.

Instead, we now replace the content with a ByteArrayContent which
contains the buffer we wrote to when calculating the size.

We implemented a new CachingByteArrayOutputStream so that we can access
the byte buffer directly rather than copying into a new byte array (for
memory performance)